# edge-e2e/wrapper/nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST API definition for End-to-end testing of the Azure IoT SDKs.\
    \  All SDK APIs that are tested by our E2E tests need to be defined in this file.\
    \  This file takes some liberties with the API definitions.  In particular, response\
    \ schemas are undefined, and error responses are also undefined."
  version: "1.0.0"
  title: "IoT SDK Device & Client REST API"
...
schemes:
- "https"
produces:
- "application/json"
paths:
...
  /device/connect/{transportType}:
    put:
      tags:
      - "device"
      summary: "Connect to the azure IoT Hub as a device"
      operationId: "deviceConnectTransportTypePUT"
      consumes:
      - "application/json"
      parameters:
      - name: "transportType"
        in: "path"
        description: "Transport to use"
        required: true
        type: "string"
        enum:
        - "amqp"
        - "amqpws"
        - "mqtt"
        - "mqttws"
        - "http"
      - name: "connectionString"
        in: "query"
        description: "connection string"
        required: true
        type: "string"
      - in: "body"
        name: "caCertificate"
        required: false
        schema:
          $ref: "#/definitions/certificate"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/connectResponse"
      x-swagger-router-controller: "Device"
  /device/{connectionId}/disconnect:
    put:
      tags:
      - "device"
      summary: "Disconnect the device"
      description: "Disconnects from Azure IoTHub service.  More specifically, closes\
        \ all connections and cleans up all resources for the active connection"
      operationId: "deviceConnectionIdDisconnectPUT"
      parameters:
      - name: "connectionId"
        in: "path"
        description: "Id for the connection"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Device"
  /device/{connectionId}/enableMethods:
    put:
      tags:
      - "device"
      summary: "Enable methods"
      operationId: "deviceConnectionIdEnableMethodsPUT"
      parameters:
      - name: "connectionId"
        in: "path"
        description: "Id for the connection"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Device"
  /device/{connectionId}/roundtripMethodCall/{methodName}:
    put:
      tags:
      - "device"
      summary: "Wait for a method call, verify the request, and return the response."
      description: "This is a workaround to deal with SDKs that only have method call\
        \ operations that are sync.  This function responds to the method with the\
        \ payload of this function, and then returns the method parameters.  Real-world\
        \ implemenatations would never do this, but this is the only same way to write\
        \ our test code right now (because the method handlers for C, Java, and probably\
        \ Python all return the method response instead of supporting an async method\
        \ call)"
      operationId: "deviceConnectionIdRoundtripMethodCallMethodNamePUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "connectionId"
        in: "path"
        description: "Id for the connection"
        required: true
        type: "string"
      - name: "methodName"
        in: "path"
        description: "name of the method to handle"
        required: true
        type: "string"
      - in: "body"
        name: "requestAndResponse"
        required: true
        schema:
          $ref: "#/definitions/roundtripMethodCallBody"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Device"
definitions:
  connectResponse:
    type: "object"
    properties:
      connectionId:
        type: "string"
        description: "unique identifier for this connection"
    description: "result of a connection to a service, device, or module client"
    example:
      connectionId: "connectionId"
  roundtripMethodCallBody:
    type: "object"
    properties:
      requestPayload:
        type: "object"
        description: "payload for the request that arrived from the service.  Used\
          \ to verify that the correct request arrived."
        properties: {}
      responsePayload:
        type: "object"
        description: "payload for the response to return to the service.  Used to\
          \ verify that the correct request arrived."
        properties: {}
      statusCode:
        type: "integer"
        description: "status code to return to the service"
    description: "parameters and response for a sync method call"
  certificate:
    type: "object"
    properties:
      cert:
        type: "string"
        description: "unique identifier for this connection"
    description: "certificate in the body of a message"